+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Retire `RSpec/InvalidPredicateMatcher` cop. ([@pirj][])
 * Remove the code responsible for filtering files to inspect. ([@pirj][])
 * Make RSpec language elements configurable. ([@sl4vr][])
+* Remove `CustomIncludeMethods` `RSpec/EmptyExampleGroup` option in favour of the new RSpec DSL configuration. ([@pirj][])
 
 ## 2.0.0.pre (2020-10-22)